replaced click with argparse, copied core.py from TLO
click is just too tedious and far more than we need for this. Replaced it with argparse

I like the idea of core with definitions for Parameters and Properties. In particular as we may wish users to register their own properties and parameters in future (there is a separate command for that)