Fix balance of sexes
We were accidentally using the probabilities for the opposite sex.
This wasn't caught before, as the numbers were very close
to 50%. This was due to another bug with calling `np.random.choice`
(we were not passing the probabilities as the right argument, so they
defaulted to uniform).
When changing to the new style, I fixed the probabilities in `choice`,
which also revealed this bug.
Have made the test stricter now that the behaviour is (hopefully!)
correct.