Test for different ratios of sexes
Also switching the use of `pytest.approx` to what I think is the
preferred usage.
I tried testing with lower ratios, but with 0.2 the result was often
slightly outside the current tolerance. It's not very realistic, but it
would be nice to know that the code works correctly for extreme cases too.